+++ src/lib/wapi/functions/delete-messages-all.js
@@ -55,6 +55,7 @@ MMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMM
 
 export async function deleteMessagesAll(chatId, messageArray, onlyLocal) {
   var userId = new Store.WidFactory.createWid(chatId);
+
   let conversation = WAPI.getChat(userId);
   if (!conversation) return false;
 
@@ -67,19 +68,33 @@ export async function deleteMessagesAll(chatId, messageArray, onlyLocal) {
       typeof msgId == 'string' ? window.Store.Msg.get(msgId) : msgId
     )
     .filter((x) => x);
-  if (messagesToDelete.length == 0) return true;
+  if (!messagesToDelete) return true;
   let jobs = onlyLocal
-    ? [conversation.sendDeleteMsgs(messagesToDelete, true)]
+    ? [
+        conversation.sendDeleteMsgs
+          ? conversation.sendDeleteMsgs(messagesToDelete, conversation)
+          : Store.WapDeleteMsg.sendDeleteMsgs(conversation, messagesToDelete)
+      ]
     : [
-        conversation.sendRevokeMsgs(
-          messagesToDelete.filter((msg) => msg.isSentByMe),
-          "Sender",
-          true
-        ),
-        conversation.sendDeleteMsgs(
-          messagesToDelete.filter((msg) => !msg.isSentByMe),
-          true
-        )
+        conversation.sendRevokeMsgs
+          ? conversation.sendRevokeMsgs(
+              messagesToDelete.filter((msg) => msg.isSentByMe),
+              conversation
+            )
+          : Store.WapDeleteMsg.sendRevokeMsgs(
+              conversation,
+              messagesToDelete.filter((msg) => msg.isSentByMe),
+              true
+            ),
+        conversation.sendDeleteMsgs
+          ? conversation.sendDeleteMsgs(
+              messagesToDelete.filter((msg) => !msg.isSentByMe),
+              conversation
+            )
+          : Store.WapDeleteMsg.sendDeleteMsgs(
+              conversation,
+              messagesToDelete.filter((msg) => !msg.isSentByMe)
+            )
       ];
   return Promise.all(jobs).then((_) => true);
 }